+++ macros/src/heap.rs
@@ -12,7 +12,7 @@ struct Input {
     heap_attrs: Vec<Attribute>,
     heap_vis: Visibility,
     heap_ident: Ident,
-    hooks: Option<[(Vec<Attribute>, ExprPath); 4]>,
+    hooks: Option<[(Vec<Attribute>, ExprPath); 2]>,
 }
 
 impl Parse for Input {
@@ -33,20 +33,7 @@ impl Parse for Input {
             input.parse::<Token![use]>()?;
             let dealloc_path = input.parse()?;
             input.parse::<Token![;]>()?;
-            let grow_in_place_attrs = input.call(Attribute::parse_outer)?;
-            input.parse::<Token![use]>()?;
-            let grow_in_place_path = input.parse()?;
-            input.parse::<Token![;]>()?;
-            let shrink_in_place_attrs = input.call(Attribute::parse_outer)?;
-            input.parse::<Token![use]>()?;
-            let shrink_in_place_path = input.parse()?;
-            input.parse::<Token![;]>()?;
-            Some([
-                (alloc_attrs, alloc_path),
-                (dealloc_attrs, dealloc_path),
-                (grow_in_place_attrs, grow_in_place_path),
-                (shrink_in_place_attrs, shrink_in_place_path),
-            ])
+            Some([(alloc_attrs, alloc_path), (dealloc_attrs, dealloc_path)])
         };
         Ok(Self { heap_attrs, heap_vis, heap_ident, hooks })
     }
@@ -73,8 +60,7 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
         pointer += pool.block * pool.capacity;
     }
     let hook_tokens = if let Some(hooks) = hooks {
-        let [(alloc_attrs, alloc_path), (dealloc_attrs, dealloc_path), (grow_in_place_attrs, grow_in_place_path), (shrink_in_place_attrs, shrink_in_place_path)] =
-            hooks;
+        let [(alloc_attrs, alloc_path), (dealloc_attrs, dealloc_path)] = hooks;
         vec![quote! {
             #(#alloc_attrs)*
             #[inline]
@@ -92,24 +78,6 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
                 pool: &::drone_core::heap::Pool,
             ) {
                 #dealloc_path(layout, pool)
-            }
-
-            #(#grow_in_place_attrs)*
-            #[inline]
-            fn grow_in_place_hook(
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) {
-                #grow_in_place_path(layout, new_size)
-            }
-
-            #(#shrink_in_place_attrs)*
-            #[inline]
-            fn shrink_in_place_hook(
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) {
-                #shrink_in_place_path(layout, new_size)
             }
         }]
     } else {
@@ -143,23 +111,15 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
                 self.pools.get_unchecked(index)
             }
 
-            #[inline]
-            unsafe fn get_pool_unchecked_mut<I>(&mut self, index: I) -> &mut I::Output
-            where
-                I: ::core::slice::SliceIndex<[::drone_core::heap::Pool]>,
-            {
-                self.pools.get_unchecked_mut(index)
-            }
-
             #(#hook_tokens)*
         }
 
         unsafe impl ::core::alloc::AllocRef for #heap_ident {
-            unsafe fn alloc(
+            fn alloc(
                 &mut self,
                 layout: ::core::alloc::Layout,
-            ) -> Result<::core::ptr::NonNull<u8>, ::core::alloc::AllocErr> {
-                ::drone_core::heap::Allocator::alloc(self, layout)
+            ) -> Result<(::core::ptr::NonNull<u8>, usize), ::core::alloc::AllocErr> {
+                ::drone_core::heap::alloc(self, layout)
             }
 
             unsafe fn dealloc(
@@ -167,83 +127,23 @@ pub fn proc_macro(input: TokenStream) -> TokenStream {
                 ptr: ::core::ptr::NonNull<u8>,
                 layout: ::core::alloc::Layout,
             ) {
-                ::drone_core::heap::Allocator::dealloc(self, ptr, layout)
-            }
-
-            fn usable_size(&self, layout: &::core::alloc::Layout) -> (usize, usize) {
-                unsafe { ::drone_core::heap::Allocator::usable_size(self, layout) }
-            }
-
-            unsafe fn realloc(
-                &mut self,
-                ptr: ::core::ptr::NonNull<u8>,
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) -> Result<::core::ptr::NonNull<u8>, ::core::alloc::AllocErr> {
-                ::drone_core::heap::Allocator::realloc(self, ptr, layout, new_size)
-            }
-
-            unsafe fn alloc_excess(
-                &mut self,
-                layout: ::core::alloc::Layout,
-            ) -> Result<::core::alloc::Excess, ::core::alloc::AllocErr> {
-                ::drone_core::heap::Allocator::alloc(self, layout)
-            }
-
-            unsafe fn realloc_excess(
-                &mut self,
-                ptr: ::core::ptr::NonNull<u8>,
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) -> Result<::core::alloc::Excess, ::core::alloc::AllocErr> {
-                ::drone_core::heap::Allocator::realloc(self, ptr, layout, new_size)
-            }
-
-            unsafe fn grow_in_place(
-                &mut self,
-                ptr: ::core::ptr::NonNull<u8>,
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) -> Result<(), ::core::alloc::CannotReallocInPlace> {
-                ::drone_core::heap::Allocator::grow_in_place(self, ptr, layout, new_size)
-            }
-
-            unsafe fn shrink_in_place(
-                &mut self,
-                ptr: ::core::ptr::NonNull<u8>,
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) -> Result<(), ::core::alloc::CannotReallocInPlace> {
-                ::drone_core::heap::Allocator::shrink_in_place(self, ptr, layout, new_size)
+                ::drone_core::heap::dealloc(self, ptr, layout)
             }
         }
 
         unsafe impl ::core::alloc::GlobalAlloc for #heap_ident {
             unsafe fn alloc(&self, layout: ::core::alloc::Layout) -> *mut u8 {
-                ::drone_core::heap::Allocator::alloc(self, layout)
-                    .map(::core::ptr::NonNull::as_ptr).unwrap_or(::core::ptr::null_mut())
+                ::drone_core::heap::alloc(self, layout)
+                    .map(|(ptr, _size)| ptr.as_ptr())
+                    .unwrap_or(::core::ptr::null_mut())
             }
 
             unsafe fn dealloc(&self, ptr: *mut u8, layout: ::core::alloc::Layout) {
-                ::drone_core::heap::Allocator::dealloc(
+                ::drone_core::heap::dealloc(
                     self,
                     ::core::ptr::NonNull::new_unchecked(ptr),
                     layout,
                 )
-            }
-
-            unsafe fn realloc(
-                &self,
-                ptr: *mut u8,
-                layout: ::core::alloc::Layout,
-                new_size: usize,
-            ) -> *mut u8 {
-                ::drone_core::heap::Allocator::realloc(
-                    self,
-                    ::core::ptr::NonNull::new_unchecked(ptr),
-                    layout,
-                    new_size,
-                ).map(::core::ptr::NonNull::as_ptr).unwrap_or(::core::ptr::null_mut())
             }
         }
     };